v1.4.5: whole-program dead-code lint, W-SYN-002, Pages workflow, CI and packaging fixes
Language and toolchain:
- Dead-code lint rewritten as whole-program analysis: references are collected
  across all project files at once and every AST node is scanned (match/decide
  arms, if-expressions, casts, closures, contracts, globals), eliminating
  false positives on cross-module projects while keeping true positives.
- W-SYN-002 (non-canonical type spellings) is now actually emitted by the parser.
- Restored documented prelude aliases and/or/xor that were accidentally dropped
  from the type checker in v1.4.5.
- Integer literals in return position now unify with the function return type,
  matching let-inference.
- dpm init <name> scaffolds <name>/ (cargo-new style); library callers passing
  a project root keep in-place behavior.
- Removed dead TryCatch plumbing from AST/DMIR/optimizer/backends (the parser
  has rejected try for a long time).
- Examples updated to canonical spellings (Str, Outcome.ok()/err()).

CI and tests:
- Fixed bridge_mmap unix-only build (mmap symbol).
- Stabilized flaky microbenchmarks (warm-up pass, min-of-runs, black_box).
- SIMD LLVM-IR test updated to the opaque-pointer emission used since v1.4.5.
- test_version_sync guards VERSION/Cargo.toml/README badges.
- Sparks seed-package tests generate their fixtures.

Docs and packaging:
- GitHub Pages workflow for the docs portal, README navigation refreshed.
- VERSION, editor extensions and npm/PyPI packages bumped to 1.4.5.

+++ src/codegen/cranelift/backend/inst_method_call.rs
@@ -298,12 +298,12 @@ pub fn compile_method_call<M: ClifModule>(
         match method {
             "sort" => {
                 let mode = ctx.builder.ins().iconst(clif_types::I64, 0);
-                let kind = ctx.builder.ins().iconst(clif_types::I64, elem_kind as i64);
+                let kind = ctx.builder.ins().iconst(clif_types::I64, elem_kind);
                 all_args.push(mode);
                 all_args.push(kind);
             }
             "remove_value" | "contains" | "index_of" => {
-                let kind = ctx.builder.ins().iconst(clif_types::I64, elem_kind as i64);
+                let kind = ctx.builder.ins().iconst(clif_types::I64, elem_kind);
                 all_args.push(kind);
             }
             _ => {}
